git-svn-id: svn://120.24.217.103/repository/outsource_projects/wu_xb/android/shop/XiaoeShop@241314 d1851369-bd24-4d56-8f42-d5f936fd0249

+++ common/src/main/java/xiaoe/com/common/utils/Dp2Px2SpUtil.java
@@ -39,9 +39,9 @@ public class Dp2Px2SpUtil {
      * @param pxVal
      * @return
      */
-    public static float px2dp(Context context, float pxVal) {
+    public static int px2dp(Context context, float pxVal) {
         final float scale = context.getResources().getDisplayMetrics().density;
-        return (pxVal / scale);
+        return (int) (pxVal / scale);
     }
 
     /**
@@ -51,7 +51,7 @@ public class Dp2Px2SpUtil {
      * @param pxVal
      * @return
      */
-    public static float px2sp(Context context, float pxVal) {
-        return (pxVal / context.getResources().getDisplayMetrics().scaledDensity);
+    public static int px2sp(Context context, float pxVal) {
+        return (int) (pxVal / context.getResources().getDisplayMetrics().scaledDensity);
     }
 }
